extract debug button logic
